Fix the problem when only one test is imported.
If you go from 0 to 1 test, we had a problem with the dropdown where you couldn't select / switch to the 1 test and get the questions to render. Now there is a default "select test" option at the top of the dropdown which re-enables switching to get out of that broken state.