Create parse_png_images.py
Adding a python script that can quickly generate a save state to upload to the tool. For example suppose you upload a cover sheet, some test pages, and then individual problems on a site like imgur and pull the HTML for the section that has the image URLs. Then you save that HTML locally and run this script so it will quickly grab the URLs for the png images and shove them into a save state that can be rendered on the site.